Add row bombs
As the name implies, these eliminate the entire row they are in (along one
direction, determined at creation). This commit also includes a new
geometry type - a hexagon with a stripe - to visualise these. These cannot
be created from special matches yet, but the level is generated with a few
of them.